requirements.txt
absl-py==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Cython==0.29.34
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
descartes==1.1.0
dill==0.3.1.1
docker-pycreds==0.4.0
entrypoints==0.4
filelock==3.4.1
fire==0.4.0
future==0.18.2
gdown==3.12.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.4
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.14.0
motmetrics==1.2.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
ncls==0.0.51
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.2.0
numpy==1.19.5
nuscenes-devkit==1.0.6
oauthlib==3.2.2
opencv-python==4.5.1.48
orjson==3.5.1
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.5
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.5
pyrsistent==0.18.0
PySocks==1.7.1
pytest==5.3.0
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==24.0.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.8.0
sentry-sdk==1.20.0
setproctitle==1.2.3
Shapely==1.8.4
six==1.16.0
smmap==5.0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.8
termcolor==1.1.0
terminado==0.13.0
terminaltables==3.1.10
testpath==0.6.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.8.0
torchaudio==0.8.0
torchsummary==1.5.1
torchvision==0.9.0
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.12
wandb==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
xmltodict==0.13.0
zipp==3.6.0
